feat: bulk action for approve / reject
