+++ src/components/ui/data-table/_components/index.ts
@@ -1,0 +1,2 @@
+export { default as FilterSelectDate } from "./filter-select-date";
+export { default as FilterStatus } from "./filter-status";
